+++ include/LIEF/DWARF/Parameter.hpp
@@ -19,6 +19,7 @@
 
 #include <memory>
 #include <string>
+#include <cstdint>
 
 namespace LIEF {
 namespace dwarf {
@@ -41,11 +42,51 @@ class LIEF_API Parameter {
     TEMPLATE_VALUE, ///< DW_TAG_template_value_parameter
     FORMAL,         ///< DW_TAG_formal_parameter
   };
+
+  /// This class exposes information about the location of a parameter
+  class LIEF_API Location {
+    public:
+    enum class Type : uint8_t {
+      UNKNOWN = 0,
+      REG,
+    };
+    Location(Type ty) :
+      type(ty)
+    {}
+
+    template<class T>
+    const T* as() const {
+      if (T::classof(this)) {
+        return static_cast<const T*>(this);
+      }
+      return nullptr;
+    }
+
+    Type type = Type::UNKNOWN;
+  };
+
+  /// This class represents a register location
+  class LIEF_API RegisterLoc : public Location {
+    public:
+    RegisterLoc(uint64_t reg_id) :
+      Location(Type::REG),
+      id(reg_id)
+    {}
+
+    static bool classof(const Location* loc) {
+      return loc->type == Type::REG;
+    }
+
+    /// DWARF id of the register
+    uint64_t id = 0;
+  };
+
   Parameter() = delete;
   Parameter(Parameter&& other);
   Parameter& operator=(Parameter&& other);
   Parameter& operator=(const Parameter&) = delete;
   Parameter(const Parameter&) = delete;
+
 
   KIND kind() const;
 
@@ -54,6 +95,10 @@ class LIEF_API Parameter {
 
   /// Type of this parameter
   std::unique_ptr<Type> type() const;
+
+  /// Location of this parameter. For instance it can be a specific register
+  /// that is not following the calling convention.
+  std::unique_ptr<Location> location() const;
 
   template<class T>
   const T* as() const {